+++ ext/sysvsem/sysvsem.c
@@ -17,7 +17,7 @@
    +----------------------------------------------------------------------+
  */
  
-/* $Id: sysvsem.c,v 1.43 2003/01/19 00:45:50 iliaa Exp $ */
+/* $Id: sysvsem.c,v 1.44 2003/03/06 23:07:27 ddhill Exp $ */
 
 /* Latest update build anc tested on Linux 2.2.14
  *
@@ -161,7 +161,7 @@ PHP_MINIT_FUNCTION(sysvsem)
    Return an id for the semaphore with the given key, and allow max_acquire (default 1) processes to acquire it simultaneously */
 PHP_FUNCTION(sem_get)
 {
-	int key, max_acquire, perm, auto_release = 1;
+	long key, max_acquire, perm, auto_release = 1;
     int semid;
 	struct sembuf sop[3];
 	int count;
